Picture_Editor_3in1/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.27.0
transformers>=4.36.0
peft>=0.8.0
accelerate>=0.25.0
datasets>=2.16.0
Pillow>=10.0.0
gradio>=4.0.0
safetensors>=0.4.0
huggingface-hub>=0.20.0
tqdm>=4.66.0
